cctv_agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
pyyaml==6.0.1
# Test dependencies
pytest==7.4.4
httpx==0.26.0
